refactor asset type handling and improve error management.

+++ BudgetFlow.Application/AssetTypes/Commands/DeleteAssetType/DeleteAssetTypeCommand.cs
@@ -1,9 +1,10 @@
 ﻿using BudgetFlow.Application.Common.Interfaces.Repositories;
+using BudgetFlow.Application.Common.Results;
 using MediatR;
 
 namespace BudgetFlow.Application.AssetTypes.Commands.DeleteAssetType
 {
-    public class DeleteAssetTypeCommand : IRequest<bool>
+    public class DeleteAssetTypeCommand : IRequest<Result<bool>>
     {
         public int ID { get; set; }
         public DeleteAssetTypeCommand(int ID)
@@ -11,19 +12,25 @@ namespace BudgetFlow.Application.AssetTypes.Commands.DeleteAssetType
             this.ID = ID;
         }
 
-        public class DeleteAssetTypeCommandHandler : IRequestHandler<DeleteAssetTypeCommand, bool>
+        public class DeleteAssetTypeCommandHandler : IRequestHandler<DeleteAssetTypeCommand, Result<bool>>
         {
             private readonly IAssetTypeRepository assetTypeRepository;
             public DeleteAssetTypeCommandHandler(IAssetTypeRepository assetTypeRepository)
             {
                 this.assetTypeRepository = assetTypeRepository;
             }
-            public Task<bool> Handle(DeleteAssetTypeCommand request, CancellationToken cancellationToken)
+            public async Task<Result<bool>> Handle(DeleteAssetTypeCommand request, CancellationToken cancellationToken)
             {
-                var result = assetTypeRepository.DeleteAssetTypeAsync(request.ID);
-                return result;
+                if (request.ID == 0)
+                    return
+                        Result.Failure<bool>("ID cannot be 0");
+
+                var deleted = await assetTypeRepository.DeleteAssetTypeAsync(request.ID);
+
+                return deleted
+                    ? Result.Success(true)
+                    : Result.Failure<bool>("Failed to delete Asset Type");
             }
         }
-
     }
 }
